refactor: change ID for an object, group services into folders
