+++ src/main/java/ispp/project/dondesiempre/services/outfits/OutfitTagService.java
@@ -1,12 +1,11 @@
-package ispp.project.dondesiempre.services;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+package ispp.project.dondesiempre.services.outfits;
 
 import ispp.project.dondesiempre.exceptions.ResourceNotFoundException;
 import ispp.project.dondesiempre.models.outfits.OutfitTag;
 import ispp.project.dondesiempre.repositories.outfits.OutfitTagRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
