+++ src/main/java/ispp/project/dondesiempre/services/products/ProductService.java
@@ -1,19 +1,18 @@
-package ispp.project.dondesiempre.services;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+package ispp.project.dondesiempre.services.products;
 
 import ispp.project.dondesiempre.exceptions.ResourceNotFoundException;
 import ispp.project.dondesiempre.models.products.Product;
 import ispp.project.dondesiempre.repositories.products.ProductRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 public class ProductService {
   private final ProductRepository productRepository;
 
-  @Transactional(readOnly = true)
+  @Transactional(readOnly = true, rollbackFor = ResourceNotFoundException.class)
   public Product findById(Integer id) throws ResourceNotFoundException {
     return productRepository
         .findById(id)
