+++ src/main/java/ispp/project/dondesiempre/services/storefronts/StorefrontService.java
@@ -1,19 +1,18 @@
-package ispp.project.dondesiempre.services;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+package ispp.project.dondesiempre.services.storefronts;
 
 import ispp.project.dondesiempre.exceptions.ResourceNotFoundException;
 import ispp.project.dondesiempre.models.storefronts.Storefront;
 import ispp.project.dondesiempre.repositories.storefronts.StorefrontRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 public class StorefrontService {
   private final StorefrontRepository storefrontRepository;
 
-  @Transactional(readOnly = true)
+  @Transactional(readOnly = true, rollbackFor = ResourceNotFoundException.class)
   public Storefront findById(Integer id) throws ResourceNotFoundException {
     return storefrontRepository
         .findById(id)
