+++ src/main/java/ispp/project/dondesiempre/services/stores/StoreService.java
@@ -1,0 +1,21 @@
+package ispp.project.dondesiempre.services.stores;
+
+import ispp.project.dondesiempre.exceptions.ResourceNotFoundException;
+import ispp.project.dondesiempre.models.stores.Store;
+import ispp.project.dondesiempre.repositories.stores.StoreRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class StoreService {
+  private final StoreRepository storeRepository;
+
+  @Transactional(readOnly = true, rollbackFor = ResourceNotFoundException.class)
+  public Store findById(Integer id) throws ResourceNotFoundException {
+    return storeRepository
+        .findById(id)
+        .orElseThrow(() -> new ResourceNotFoundException("Store with ID " + id + " not found."));
+  }
+}
